Görüntü_On_Isleme/requirements.txt
numpy
pillow
matplotlib
scikit-image
opencv-python
